Moved convertToFreq and convertToVel into Noise model as they are coded especially for NoiseModel
